better way to make the logo + more variations
